+++ src/Core/src/Sts/StsClient.php
@@ -97,9 +97,7 @@ class StsClient extends AbstractApi
                 'endpoint' => 'https://sts.amazonaws.com',
                 'signRegion' => 'us-east-1',
                 'signService' => 'sts',
-                'signVersions' => [
-                    0 => 'v4',
-                ],
+                'signVersions' => ['v4'],
             ];
         }
 
@@ -125,86 +123,68 @@ class StsClient extends AbstractApi
             case 'us-west-1':
             case 'us-west-2':
                 return [
-                    'endpoint' => 'https://sts.%region%.amazonaws.com',
+                    'endpoint' => "https://sts.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'cn-north-1':
             case 'cn-northwest-1':
                 return [
-                    'endpoint' => 'https://sts.%region%.amazonaws.com.cn',
+                    'endpoint' => "https://sts.$region.amazonaws.com.cn",
                     'signRegion' => $region,
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-gov-east-1':
             case 'us-gov-west-1':
                 return [
-                    'endpoint' => 'https://sts.%region%.amazonaws.com',
+                    'endpoint' => "https://sts.$region.amazonaws.com",
                     'signRegion' => $region,
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-iso-east-1':
                 return [
-                    'endpoint' => 'https://sts.%region%.c2s.ic.gov',
+                    'endpoint' => "https://sts.$region.c2s.ic.gov",
                     'signRegion' => $region,
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-isob-east-1':
                 return [
-                    'endpoint' => 'https://sts.%region%.sc2s.sgov.gov',
+                    'endpoint' => "https://sts.$region.sc2s.sgov.gov",
                     'signRegion' => $region,
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-1-fips':
                 return [
                     'endpoint' => 'https://sts-fips.us-east-1.amazonaws.com',
                     'signRegion' => 'us-east-1',
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-east-2-fips':
                 return [
                     'endpoint' => 'https://sts-fips.us-east-2.amazonaws.com',
                     'signRegion' => 'us-east-2',
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-1-fips':
                 return [
                     'endpoint' => 'https://sts-fips.us-west-1.amazonaws.com',
                     'signRegion' => 'us-west-1',
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
             case 'us-west-2-fips':
                 return [
                     'endpoint' => 'https://sts-fips.us-west-2.amazonaws.com',
                     'signRegion' => 'us-west-2',
                     'signService' => 'sts',
-                    'signVersions' => [
-                        0 => 'v4',
-                    ],
+                    'signVersions' => ['v4'],
                 ];
         }
 
@@ -213,16 +193,22 @@ class StsClient extends AbstractApi
 
     protected function getServiceCode(): string
     {
+        @trigger_error('Using the client with an old version of Core is deprecated. Run "composer update async-aws/core".', \E_USER_DEPRECATED);
+
         return 'sts';
     }
 
     protected function getSignatureScopeName(): string
     {
+        @trigger_error('Using the client with an old version of Core is deprecated. Run "composer update async-aws/core".', \E_USER_DEPRECATED);
+
         return 'sts';
     }
 
     protected function getSignatureVersion(): string
     {
+        @trigger_error('Using the client with an old version of Core is deprecated. Run "composer update async-aws/core".', \E_USER_DEPRECATED);
+
         return 'v4';
     }
 }