+++ src/Service/S3/src/Input/CompleteMultipartUploadRequest.php
@@ -136,7 +136,7 @@ final class CompleteMultipartUploadRequest extends Input
             throw new InvalidArgument(sprintf('Missing parameter "Key" for "%s". The value cannot be null.', __CLASS__));
         }
         $uri['Key'] = $v;
-        $uriString = "/{$uri['Bucket']}/{$uri['Key']}";
+        $uriString = '/' . rawurlencode($uri['Bucket']) . '/' . str_replace('%2F', '/', rawurlencode($uri['Key']));
 
         // Prepare Body