+++ src/Service/S3/src/Input/CreateMultipartUploadRequest.php
@@ -555,7 +555,7 @@ final class CreateMultipartUploadRequest extends Input
             throw new InvalidArgument(sprintf('Missing parameter "Key" for "%s". The value cannot be null.', __CLASS__));
         }
         $uri['Key'] = $v;
-        $uriString = "/{$uri['Bucket']}/{$uri['Key']}?uploads";
+        $uriString = '/' . rawurlencode($uri['Bucket']) . '/' . str_replace('%2F', '/', rawurlencode($uri['Key'])) . '?uploads';
 
         // Prepare Body
         $body = '';