+++ src/Service/S3/src/Input/DeleteObjectsRequest.php
@@ -135,7 +135,7 @@ final class DeleteObjectsRequest extends Input
             throw new InvalidArgument(sprintf('Missing parameter "Bucket" for "%s". The value cannot be null.', __CLASS__));
         }
         $uri['Bucket'] = $v;
-        $uriString = "/{$uri['Bucket']}?delete";
+        $uriString = '/' . rawurlencode($uri['Bucket']) . '?delete';
 
         // Prepare Body