+++ src/Service/DynamoDb/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## NOT RELEASED
+
+### Added
+
+- Added operation `UpdateTimeToLiveInput`
+
+### Deprecation
+
+- Removes methods `getServiceCode`, `getSignatureVersion` and `getSignatureScopeName` from Client.
+
 ## 0.2.1
 
 ### Fixed