+++ src/Service/Lambda/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## NOT RELEASED
+
+### Deprecation
+
+- Removes methods `getServiceCode`, `getSignatureVersion` and `getSignatureScopeName` from Client.
+
+### Fixed
+
+- Fixed issue when Layer, Function or Version contained a special char `#`
+
 ## 0.4.2
 
 ### Fixed